requirements.txt
anytree==2.8.0
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
Django==4.1.4
djangorestframework==3.14.0
et-xmlfile==1.1.0
gitdb==4.0.10
GitPython==3.1.29
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
MarkupSafe==2.1.1
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.2
psycopg2-binary==2.9.5
pycparser==2.21
pycryptodome==3.16.0
PyJWT==2.6.0
pyotp==2.7.0
python-dateutil==2.8.2
pytz==2022.6
qrcode==7.3.1
requests==2.28.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.44
sqlcipher3==0.4.8
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
validate-email==1.3
Werkzeug==2.2.2
wrapt==1.14.1
xlrd==2.0.1
